+++ Backend/order/src/main/java/com/urbanstyle/order/Repository/UserOrderProductRepository.java
@@ -47,7 +47,7 @@ public interface UserOrderProductRepository extends PagingAndSortingRepository<U
 	UserOrderProducts findByIdAndVendorId(long orderProductId, long vendorId);
 
 	@Query("Select  uop from UserOrderProducts uop where uop.vendor.id =?2 and uop.status=?1 order by uop.userOrder.orderPlacedDate desc")
-	List<UserOrderProducts> getAllOrderByStatusAndUserId(String status, int vendorId, Pageable pagable);
+	List<UserOrderProducts> getAllOrderByStatusAndUserId(String status, long vendorId, Pageable pagable);
 
 	long countByUserOrderUserId(long userId);
